serious refactoring of tests
