+++ sql/core/src/test/scala/org/apache/spark/sql/sources/v2/DataSourceV2DataFrameSuite.scala
@@ -17,13 +17,10 @@
 
 package org.apache.spark.sql.sources.v2
 
-import org.scalatest.BeforeAndAfter
+import org.apache.spark.sql.{DataFrame, Row, SaveMode}
 
-import org.apache.spark.sql.{QueryTest, Row}
-import org.apache.spark.sql.internal.SQLConf.{PARTITION_OVERWRITE_MODE, PartitionOverwriteMode}
-import org.apache.spark.sql.test.SharedSparkSession
-
-class DataSourceV2DataFrameSuite extends QueryTest with SharedSparkSession with BeforeAndAfter {
+class DataSourceV2DataFrameSuite
+  extends InsertIntoTests(supportsDynamicOverwrite = true, includeSQLTests = false) {
   import testImplicits._
 
   before {
@@ -31,25 +28,19 @@ class DataSourceV2DataFrameSuite extends QueryTest with SharedSparkSession with 
     spark.conf.set("spark.sql.catalog.testcat2", classOf[TestInMemoryTableCatalog].getName)
   }
 
-  test("insertInto: append") {
-    val t1 = "testcat.ns1.ns2.tbl"
-    withTable(t1) {
-      sql(s"CREATE TABLE $t1 (id bigint, data string) USING foo")
-      val df = Seq((1L, "a"), (2L, "b"), (3L, "c")).toDF("id", "data")
-      df.write.insertInto(t1)
-      checkAnswer(spark.table(t1), df)
-    }
+  override protected val catalogAndNamespace: String = "testcat.ns1.ns2.tbls"
+  override protected val v2Format: String = classOf[FakeV2Provider].getName
+
+  override def verifyTable(tableName: String, expected: DataFrame): Unit = {
+    checkAnswer(spark.table(tableName), expected)
   }
 
-  test("insertInto: append by position") {
-    val t1 = "testcat.ns1.ns2.tbl"
-    withTable(t1) {
-      sql(s"CREATE TABLE $t1 (id bigint, data string) USING foo")
-      val df = Seq((1L, "a"), (2L, "b"), (3L, "c")).toDF("id", "data")
-      val dfr = Seq((1L, "a"), (2L, "b"), (3L, "c")).toDF("data", "id")
-      dfr.write.insertInto(t1)
-      checkAnswer(spark.table(t1), df)
+  override protected def doInsert(tableName: String, insert: DataFrame, mode: SaveMode): Unit = {
+    val dfw = insert.write.format(v2Format)
+    if (mode != null) {
+      dfw.mode(mode)
     }
+    dfw.insertInto(tableName)
   }
 
   test("insertInto: append across catalog") {
@@ -62,83 +53,6 @@ class DataSourceV2DataFrameSuite extends QueryTest with SharedSparkSession with 
       df.write.insertInto(t1)
       spark.table(t1).write.insertInto(t2)
       checkAnswer(spark.table(t2), df)
-    }
-  }
-
-  test("insertInto: append partitioned table") {
-    val t1 = "testcat.ns1.ns2.tbl"
-    withTable(t1) {
-      sql(s"CREATE TABLE $t1 (id bigint, data string) USING foo PARTITIONED BY (id)")
-      val df = Seq((1L, "a"), (2L, "b"), (3L, "c")).toDF("id", "data")
-      df.write.insertInto(t1)
-      checkAnswer(spark.table(t1), df)
-    }
-  }
-
-  test("insertInto: overwrite non-partitioned table") {
-    val t1 = "testcat.ns1.ns2.tbl"
-    withTable(t1) {
-      sql(s"CREATE TABLE $t1 (id bigint, data string) USING foo")
-      val df = Seq((1L, "a"), (2L, "b"), (3L, "c")).toDF("id", "data")
-      val df2 = Seq((4L, "d"), (5L, "e"), (6L, "f")).toDF("id", "data")
-      df.write.insertInto(t1)
-      df2.write.mode("overwrite").insertInto(t1)
-      checkAnswer(spark.table(t1), df2)
-    }
-  }
-
-  test("insertInto: overwrite partitioned table in static mode") {
-    withSQLConf(PARTITION_OVERWRITE_MODE.key -> PartitionOverwriteMode.STATIC.toString) {
-      val t1 = "testcat.ns1.ns2.tbl"
-      withTable(t1) {
-        sql(s"CREATE TABLE $t1 (id bigint, data string) USING foo PARTITIONED BY (id)")
-        Seq((2L, "dummy"), (4L, "keep")).toDF("id", "data").write.insertInto(t1)
-        val df = Seq((1L, "a"), (2L, "b"), (3L, "c")).toDF("id", "data")
-        df.write.mode("overwrite").insertInto(t1)
-        checkAnswer(spark.table(t1), df)
-      }
-    }
-  }
-
-
-  test("insertInto: overwrite partitioned table in static mode by position") {
-    withSQLConf(PARTITION_OVERWRITE_MODE.key -> PartitionOverwriteMode.STATIC.toString) {
-      val t1 = "testcat.ns1.ns2.tbl"
-      withTable(t1) {
-        sql(s"CREATE TABLE $t1 (id bigint, data string) USING foo PARTITIONED BY (id)")
-        Seq((2L, "dummy"), (4L, "keep")).toDF("id", "data").write.insertInto(t1)
-        val df = Seq((1L, "a"), (2L, "b"), (3L, "c")).toDF("id", "data")
-        val dfr = Seq((1L, "a"), (2L, "b"), (3L, "c")).toDF("data", "id")
-        dfr.write.mode("overwrite").insertInto(t1)
-        checkAnswer(spark.table(t1), df)
-      }
-    }
-  }
-
-  test("insertInto: overwrite partitioned table in dynamic mode") {
-    withSQLConf(PARTITION_OVERWRITE_MODE.key -> PartitionOverwriteMode.DYNAMIC.toString) {
-      val t1 = "testcat.ns1.ns2.tbl"
-      withTable(t1) {
-        sql(s"CREATE TABLE $t1 (id bigint, data string) USING foo PARTITIONED BY (id)")
-        Seq((2L, "dummy"), (4L, "keep")).toDF("id", "data").write.insertInto(t1)
-        val df = Seq((1L, "a"), (2L, "b"), (3L, "c")).toDF("id", "data")
-        df.write.mode("overwrite").insertInto(t1)
-        checkAnswer(spark.table(t1), df.union(sql("SELECT 4L, 'keep'")))
-      }
-    }
-  }
-
-  test("insertInto: overwrite partitioned table in dynamic mode by position") {
-    withSQLConf(PARTITION_OVERWRITE_MODE.key -> PartitionOverwriteMode.DYNAMIC.toString) {
-      val t1 = "testcat.ns1.ns2.tbl"
-      withTable(t1) {
-        sql(s"CREATE TABLE $t1 (id bigint, data string) USING foo PARTITIONED BY (id)")
-        Seq((2L, "dummy"), (4L, "keep")).toDF("id", "data").write.insertInto(t1)
-        val df = Seq((1L, "a"), (2L, "b"), (3L, "c")).toDF("id", "data")
-        val dfr = Seq((1L, "a"), (2L, "b"), (3L, "c")).toDF("data", "id")
-        dfr.write.mode("overwrite").insertInto(t1)
-        checkAnswer(spark.table(t1), df.union(sql("SELECT 4L, 'keep'")))
-      }
     }
   }
